refac: refatora para usar uma consulta JPQL para o endpoint de buscar episodios por temporada

+++ 3356-java-screenmatch-web-main/src/main/java/br/com/alura/screenmatch/controller/SerieController.java
@@ -41,7 +41,7 @@ public class SerieController {
     }
 
     @GetMapping("/{id}/temporadas/{numero}")
-    public List<EpisodioDTO> obterTemporadasPorNumero(@PathVariable Long id, @PathVariable Integer numero) {
+    public List<EpisodioDTO> obterTemporadasPorNumero(@PathVariable Long id, @PathVariable Long numero) {
         return serieService.obterTemporadasPorNumero(id, numero);
     }
 }
